+++ decomp/ST.exe/functions/0061F290/decomp.c
@@ -71,7 +71,7 @@ undefined4 __thiscall STLightC::sub_0061F290(STLightC *this,int param_1)
                     (VisibleClassTy::sub_00558C00
                                (this_00,this_00->field_010C,iVar3,local_8,&local_18,&local_1c),
                     iVar4 < 0)) || ((4 < iVar4 || (local_18 < 0)))) ||
-                  ((this_00->field_0030 <= local_18 ||
+                  (((int)this_00->field_0030 <= local_18 ||
                    ((iVar4 = g_centeredOffsets5[iVar4] + local_1c, iVar4 < 0 ||
                     (this_00->field_0034 <= iVar4)))))) ||
                  ((this_00->field_004C == nullptr ||
