+++ decomp/ST.exe/functions/00635BF0/decomp.c
@@ -56,7 +56,7 @@ undefined4 FUN_00635bf0(int param_1,int param_2,int param_3)
                      (g_visibleClass_00802A88,g_visibleClass_00802A88->field_010C,iVar4,iVar5,
                       &param_2,&param_3), uVar3 = param_1, -1 < param_1)))) &&
         ((param_1 < 5 &&
-         ((((-1 < param_2 && (param_2 < pVVar2->field_0030)) &&
+         ((((-1 < param_2 && (param_2 < (int)pVVar2->field_0030)) &&
            (uVar3 = g_centeredOffsets5[param_1] + param_3, -1 < (int)uVar3)) &&
           (((int)uVar3 < pVVar2->field_0034 && (pVVar2->field_004C != nullptr)))))))) &&
        (uVar3 = (uint)pVVar2->field_004C[param_2 + uVar3 * pVVar2->field_0030], uVar3 == 0)) {
