+++ decomp/ST.exe/functions/0063C510/decomp.c
@@ -83,7 +83,7 @@ int __thiscall FUN_0063c510(void *this,int param_1)
               (VisibleClassTy::sub_00558C00
                          (g_visibleClass_00802A88,g_visibleClass_00802A88->field_010C,iVar3,iVar4,
                           &local_c,&local_10), local_8 < 0)) || ((4 < local_8 || (local_c < 0)))) ||
-            ((pVVar2->field_0030 <= local_c ||
+            (((int)pVVar2->field_0030 <= local_c ||
              ((local_10 = g_centeredOffsets5[local_8] + local_10, local_10 < 0 ||
               (pVVar2->field_0034 <= local_10)))))) ||
            ((pVVar2->field_004C == nullptr ||
