+++ decomp/ST.exe/functions/00643E20/decomp.c
@@ -59,7 +59,7 @@ undefined4 __fastcall FUN_00643e20(AnonShape_00643E20_B7FEAA75 *param_1,undefine
                      (g_visibleClass_00802A88,g_visibleClass_00802A88->field_010C,iVar5,iVar8,
                       &local_1c,&local_20), iVar9 < 0)) ||
          (((4 < iVar9 || (local_1c < 0)) ||
-          ((pVVar3->field_0030 <= local_1c ||
+          (((int)pVVar3->field_0030 <= local_1c ||
            ((local_20 = g_centeredOffsets5[iVar9] + local_20, local_20 < 0 ||
             (pVVar3->field_0034 <= local_20)))))))) || (pVVar3->field_004C == nullptr)) ||
        (pVVar3->field_004C[local_1c + local_20 * pVVar3->field_0030] != 0)) {
